[hdpowerview] refactor unit tests into two classes
