+++ bundles/org.openhab.binding.hdpowerview/src/test/java/org/openhab/binding/hdpowerview/HDPowerViewJUnitTests.java
@@ -57,8 +57,6 @@ public class HDPowerViewJUnitTests {
     private static final Pattern VALID_IP_V4_ADDRESS = Pattern
             .compile("\\b((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)(\\.|$)){4}\\b");
 
-    private final ShadeCapabilitiesDatabase db = new ShadeCapabilitiesDatabase();
-
     /*
      * load a test JSON string from a file.
      */
@@ -190,7 +188,8 @@ public class HDPowerViewJUnitTests {
                     assertNotNull(capabilitiesValue);
 
                     if (positions != null && capabilitiesValue != null) {
-                        Capabilities capabilities = db.getCapabilities(capabilitiesValue.intValue());
+                        Capabilities capabilities = new ShadeCapabilitiesDatabase()
+                                .getCapabilities(capabilitiesValue.intValue());
 
                         State pos = positions.getState(capabilities, PRIMARY_POSITION);
                         assertEquals(PercentType.class, pos.getClass());
@@ -248,73 +247,6 @@ public class HDPowerViewJUnitTests {
                 }
             }
         }
-    }
-
-    /**
-     * Test parsing of ShadePosition (shade fully up).
-     *
-     */
-    @Test
-    public void testShadePositionParsingFullyUp() {
-        Capabilities capabilities = db.getCapabilities(0);
-        ShadePosition test = new ShadePosition().setPosition(capabilities, PRIMARY_POSITION, 0);
-        assertNotNull(test);
-        State pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(0, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, VANE_TILT_POSITION);
-        assertTrue(UnDefType.UNDEF.equals(pos));
-    }
-
-    /**
-     * Test parsing of ShadePosition (shade fully down (method 1)).
-     *
-     */
-    @Test
-    public void testShadePositionParsingShadeFullyDown1() {
-        Capabilities capabilities = db.getCapabilities(0);
-        ShadePosition test = new ShadePosition().setPosition(capabilities, PRIMARY_POSITION, 100);
-        assertNotNull(test);
-        State pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, VANE_TILT_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(0, ((PercentType) pos).intValue());
-    }
-
-    /**
-     * Test parsing of ShadePosition (shade fully down (method 2)).
-     *
-     */
-    @Test
-    public void testShadePositionParsingShadeFullyDown2() {
-        Capabilities capabilities = db.getCapabilities(0);
-        ShadePosition test = new ShadePosition().setPosition(capabilities, VANE_TILT_POSITION, 0);
-        assertNotNull(test);
-        State pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, VANE_TILT_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(0, ((PercentType) pos).intValue());
-    }
-
-    /**
-     * Test parsing of ShadePosition (shade fully down (method 2) and vane fully open).
-     *
-     */
-    @Test
-    public void testShadePositionParsingShadeFullyDownVaneOpen() {
-        Capabilities capabilities = db.getCapabilities(0);
-        ShadePosition test = new ShadePosition().setPosition(capabilities, VANE_TILT_POSITION, 100);
-        assertNotNull(test);
-        State pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, VANE_TILT_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
     }
 
     /**
@@ -409,7 +341,7 @@ public class HDPowerViewJUnitTests {
                     assertNotNull(capabilitiesValue);
                     if (capabilitiesValue != null) {
                         assertEquals(7, capabilitiesValue.intValue());
-
+                        ShadeCapabilitiesDatabase db = new ShadeCapabilitiesDatabase();
                         Capabilities capabilities = db.getCapabilities(capabilitiesValue);
 
                         State pos = shadePos.getState(capabilities, PRIMARY_POSITION);
@@ -463,240 +395,5 @@ public class HDPowerViewJUnitTests {
                 }
             }
         }
-    }
-
-    /**
-     * General tests of the database of known types.
-     */
-    @Test
-    public void testKnownTypesDatabase() {
-        assertTrue(db.isTypeInDatabase(4));
-        assertTrue(db.isCapabilitiesInDatabase(0));
-
-        assertTrue(db.getCapabilities(6).isPrimaryStateInverted());
-        assertTrue(db.getCapabilities(7).supportsSecondary());
-
-        assertEquals(db.getType(4).getCapabilities(), 0);
-        assertEquals(db.getType(-1).getCapabilities(), -1);
-
-        assertFalse(db.isTypeInDatabase(99));
-        assertFalse(db.isCapabilitiesInDatabase(99));
-
-        assertFalse(db.getCapabilities(0).isPrimaryStateInverted());
-        assertFalse(db.getCapabilities(-1).isPrimaryStateInverted());
-        assertFalse(db.getCapabilities(99).isPrimaryStateInverted());
-
-        assertFalse(db.getCapabilities(0).supportsSecondary());
-        assertFalse(db.getCapabilities(-1).supportsSecondary());
-        assertFalse(db.getCapabilities(99).supportsSecondary());
-    }
-
-    /**
-     * On dual rail shades, it should not be possible to drive the upper rail below the lower rail, or vice-versa. So
-     * the binding code applies constraints on setting such positions. This test checks that the constraint code is
-     * working.
-     */
-    @Test
-    public void testDualRailConstraints() {
-        ShadePosition shade = new ShadePosition();
-        Capabilities caps = db.getCapabilities(7);
-
-        // ==== OK !! primary at bottom, secondary at top ====
-        shade.setPosition(caps, PRIMARY_POSITION, 100).setPosition(caps, SECONDARY_POSITION, 0);
-        assertEquals(PercentType.HUNDRED, shade.getState(caps, PRIMARY_POSITION));
-        assertEquals(PercentType.ZERO, shade.getState(caps, SECONDARY_POSITION));
-
-        // ==== OK !! primary at middle, secondary at top ====
-        shade.setPosition(caps, PRIMARY_POSITION, 50).setPosition(caps, SECONDARY_POSITION, 0);
-        assertEquals(new PercentType(50), shade.getState(caps, PRIMARY_POSITION));
-        assertEquals(PercentType.ZERO, shade.getState(caps, SECONDARY_POSITION));
-
-        // ==== OK !! primary at middle, secondary at middle ====
-        shade.setPosition(caps, PRIMARY_POSITION, 50).setPosition(caps, SECONDARY_POSITION, 50);
-        assertEquals(new PercentType(50), shade.getState(caps, PRIMARY_POSITION));
-        assertEquals(new PercentType(50), shade.getState(caps, SECONDARY_POSITION));
-
-        // ==== IMPOSSIBLE !! secondary at middle, primary above => test the constraining code ====
-        shade.setPosition(caps, SECONDARY_POSITION, 0).setPosition(caps, PRIMARY_POSITION, 100);
-        shade.setPosition(caps, SECONDARY_POSITION, 40).setPosition(caps, PRIMARY_POSITION, 25);
-        assertEquals(new PercentType(40), shade.getState(caps, SECONDARY_POSITION));
-        assertEquals(new PercentType(40), shade.getState(caps, PRIMARY_POSITION));
-
-        // ==== OK !! secondary at middle, primary below ====
-        shade.setPosition(caps, SECONDARY_POSITION, 0).setPosition(caps, PRIMARY_POSITION, 100);
-        shade.setPosition(caps, SECONDARY_POSITION, 50).setPosition(caps, PRIMARY_POSITION, 75);
-        assertEquals(new PercentType(50), shade.getState(caps, SECONDARY_POSITION));
-        assertEquals(new PercentType(75), shade.getState(caps, PRIMARY_POSITION));
-
-        // ==== IMPOSSIBLE !! primary at middle, secondary below => test the constraining code ====
-        shade.setPosition(caps, SECONDARY_POSITION, 0).setPosition(caps, PRIMARY_POSITION, 100);
-        shade.setPosition(caps, PRIMARY_POSITION, 60).setPosition(caps, SECONDARY_POSITION, 75);
-        assertEquals(new PercentType(60), shade.getState(caps, PRIMARY_POSITION));
-        assertEquals(new PercentType(60), shade.getState(caps, SECONDARY_POSITION));
-
-        // ==== OK !! primary at middle, secondary above ====
-        shade.setPosition(caps, SECONDARY_POSITION, 0).setPosition(caps, PRIMARY_POSITION, 100);
-        shade.setPosition(caps, PRIMARY_POSITION, 60).setPosition(caps, SECONDARY_POSITION, 25);
-        assertEquals(new PercentType(60), shade.getState(caps, PRIMARY_POSITION));
-        assertEquals(new PercentType(25), shade.getState(caps, SECONDARY_POSITION));
-    }
-
-    /**
-     * Test parsing of DuoLite shades having a secondary blackout shade.
-     *
-     */
-    @Test
-    public void testDuoliteShadePositionParsing() {
-        ShadePosition test;
-        State pos;
-
-        // blackout shades have capabilities 8
-        Capabilities capabilities = db.getCapabilities(8);
-
-        // both shades up
-        test = new ShadePosition().setPosition(capabilities, PRIMARY_POSITION, 0);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(0, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertTrue(PercentType.ZERO.equals(pos));
-
-        // front shade 50% down
-        test = new ShadePosition().setPosition(capabilities, PRIMARY_POSITION, 50);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(50, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertTrue(PercentType.ZERO.equals(pos));
-
-        // front shade 100% down, back shade 0% down
-        test = new ShadePosition().setPosition(capabilities, PRIMARY_POSITION, 100);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertTrue(PercentType.ZERO.equals(pos));
-
-        // front shade 100% down, back shade 0% down (ALTERNATE)
-        test = new ShadePosition().setPosition(capabilities, SECONDARY_POSITION, 0);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertTrue(PercentType.ZERO.equals(pos));
-
-        // front shade 100% down, back shade 50% down
-        test = new ShadePosition().setPosition(capabilities, SECONDARY_POSITION, 50);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(50, ((PercentType) pos).intValue());
-
-        // front shade 100% down, back shade 100% down
-        test = new ShadePosition().setPosition(capabilities, SECONDARY_POSITION, 100);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-    }
-
-    /**
-     * Test parsing of DuoLite shades having both a secondary blackout shade, and tilt anywhere functionality.
-     *
-     */
-    @Test
-    public void testDuoliteTiltShadePositionParsing() {
-        ShadePosition test;
-        State pos;
-
-        // blackout shades with tilt have capabilities 9
-        Capabilities capabilities = db.getCapabilities(9);
-
-        // both shades up, tilt 0%
-        test = new ShadePosition().setPosition(capabilities, PRIMARY_POSITION, 0).setPosition(capabilities,
-                VANE_TILT_POSITION, 0);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(0, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertTrue(PercentType.ZERO.equals(pos));
-
-        // front shade 50% down, tilt 30%
-        test = new ShadePosition().setPosition(capabilities, PRIMARY_POSITION, 50).setPosition(capabilities,
-                VANE_TILT_POSITION, 30);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(50, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertTrue(PercentType.ZERO.equals(pos));
-        pos = test.getState(capabilities, VANE_TILT_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(30, ((PercentType) pos).intValue());
-
-        // front shade 100% down, back shade 0% down, tilt 30%
-        test = new ShadePosition().setPosition(capabilities, PRIMARY_POSITION, 100).setPosition(capabilities,
-                VANE_TILT_POSITION, 30);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertTrue(PercentType.ZERO.equals(pos));
-        pos = test.getState(capabilities, VANE_TILT_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(30, ((PercentType) pos).intValue());
-
-        // front shade 100% down, back shade 0% down, tilt 30% (ALTERNATE)
-        test = new ShadePosition().setPosition(capabilities, SECONDARY_POSITION, 0).setPosition(capabilities,
-                VANE_TILT_POSITION, 30);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertTrue(PercentType.ZERO.equals(pos));
-        pos = test.getState(capabilities, VANE_TILT_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(30, ((PercentType) pos).intValue());
-
-        // front shade 100% down, back shade 50% down, tilt 30%
-        test = new ShadePosition().setPosition(capabilities, SECONDARY_POSITION, 50).setPosition(capabilities,
-                VANE_TILT_POSITION, 30);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(50, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, VANE_TILT_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(30, ((PercentType) pos).intValue());
-
-        // front shade 100% down, back shade 100% down, tilt 70%
-        test = new ShadePosition().setPosition(capabilities, SECONDARY_POSITION, 100).setPosition(capabilities,
-                VANE_TILT_POSITION, 70);
-        assertNotNull(test);
-        pos = test.getState(capabilities, PRIMARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, SECONDARY_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(100, ((PercentType) pos).intValue());
-        pos = test.getState(capabilities, VANE_TILT_POSITION);
-        assertEquals(PercentType.class, pos.getClass());
-        assertEquals(70, ((PercentType) pos).intValue());
     }
 }
